Extract home menu handling from ConnectionState to new class
